requirements.txt
typer
fastapi
uvicorn
sqlalchemy
pandas
PyMySQL
mysql-connector-python
pyodbc
python-dotenv
pyyaml
openpyxl

